Add iTerm2 inline-image terminal backend (--iterm)
Render the game to the terminal via iTerm2's OSC 1337 inline-image
protocol, alongside the existing sixel backend. Each frame is expanded to
RGB888, PNG-encoded with the vendored stb_image_write (its bundled deflate
means no new zlib dependency), base64-encoded, and emitted as a
File=inline=1 sequence. Unlike sixel/kitty this renders in VS Code's
integrated terminal (xterm.js image addon), as well as iTerm2 and WezTerm,
and gives full 24-bit colour with no palette banding.

The sixel and iTerm2 backends differ only in how a finished frame becomes
bytes, so the shared machinery (raw-mode input, monotonic tick/delay
source, alternate-screen handling, RGSS::Input forwarding) is extracted
from sixel.cxx into a new terminal.cxx / terminal.hxx. Each protocol is now
just an encoder plus a thin *_display_create wrapper over
terminal_display_create.

- New --iterm / --iterm_scale flags in main.cxx; --sixel and --iterm are
  mutually exclusive.
- ADR 0002 documents the rationale and bandwidth trade-off (PNG compresses
  flat tile/sprite regions well but spends CPU per frame).
